requirements.txt
langchain
langchain_community
langchain_ollama
requests
pypdf
faiss-cpu
tiktoken
streamlit
python-dotenv
